Refactor logging decorator to a separate file and update import statement in auth.py
